signet: roll back ComputeModifiedMerkleRoot() to upstream version
In 4e84fe667431 we moved to a simpler way of computing the MerkleRoot. Thus, we
no longer need to expose ComputeModifiedMerkleRoot() externally, and there is no
need to the additional "mutated" parameter.

This change reverts every modifications we have done to
ComputeModifiedMerkleRoot(), which is now the same as in bitcoin.