build: disable "-werror" while we stay on bitcoin v23.x
Bitcoin v23.0 generates warnings with gcc 12, and does not copile at all on gcc
13. While the second problem will be solved updating to bitcoin v23.2, the
problem of triggering many warnings will not go away while we stay on v23.

This change relaxes the compilation flags in order to tolerate warnings, with
the aim of restoring it as soon as we move to a more recent major version.

Without this change it would not be possible, for example, to compile
itcoin-core on Fedora 37 (gcc 12).