ci: execute "apt update" before the first "apt install"
Apparently, the base image of the Github CI runner is not updated very often,
and it does not keep up with the evolving list of mirrors.

The "apt update" was not there on purpose, but but maybe it is more pragmatic to
just add it, otherwise the CI would start to fail due to random 404s when
installing packages.